chore: make clippy happy
